backend/pyproject.toml
[project]
name = "auroramind-backend"
version = "0.1.0"
description = "AuroraMind Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0,<5",
    "fastapi>=0.121.2",
    "passlib>=1.7.4",
    "psycopg[binary]>=3.2.13",
    "pydantic[email]>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "greenlet>=3.2.4",
]
